Revert "install v2"
This reverts commit 6f7c12d59da7c45a56bfa994a565564d6c79992d.
