+++ pkg/appinstaller/helm.go
@@ -8,6 +8,8 @@ import (
 	"fmt"
 	"net/http"
 	"net/http/httputil"
+	"os"
+	"path/filepath"
 	"strconv"
 	"time"
 
@@ -25,6 +27,7 @@ import (
 	"bytetrade.io/web3os/app-service/pkg/kubesphere"
 	"bytetrade.io/web3os/app-service/pkg/tapr"
 	"bytetrade.io/web3os/app-service/pkg/users/userspace"
+	userspacev1 "bytetrade.io/web3os/app-service/pkg/users/userspace/v1"
 	"bytetrade.io/web3os/app-service/pkg/utils"
 	apputils "bytetrade.io/web3os/app-service/pkg/utils/app"
 
@@ -39,6 +42,7 @@ import (
 	kerrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 	"k8s.io/klog/v2"
@@ -89,12 +93,189 @@ type HelmOps struct {
 	//dyClient dynamic.Interface
 	client  *clientset.ClientSet
 	options Opt
-
-	installStrategy InstallStrategy
 }
 
-func (h *HelmOps) setInstallStrategy(strategy InstallStrategy) {
-	h.installStrategy = strategy
+func (h *HelmOps) install(values map[string]interface{}) error {
+	_, err := h.status()
+	if err == nil {
+		return driver.ErrReleaseExists
+	}
+	if errors.Is(err, driver.ErrReleaseNotFound) {
+		return helm.InstallCharts(h.ctx, h.actionConfig, h.settings, h.app.AppName, h.app.ChartsName, h.app.RepoURL, h.app.Namespace, values)
+	}
+	return err
+}
+
+// Install makes install operation for an application.
+func (h *HelmOps) Install_deprecated() error {
+	values, err := h.setValues()
+	if err != nil {
+		return err
+	}
+	namespace := fmt.Sprintf("%s-%s", "user-system", h.app.OwnerName)
+	if err := tapr.Apply(h.app.Middleware, h.kubeConfig, h.app.AppName, h.app.Namespace,
+		namespace, h.token, h.app.ChartsName, h.app.OwnerName, values); err != nil {
+		klog.Errorf("Failed to apply middleware err=%v", err)
+		return err
+	}
+	err = h.install(values)
+	if err != nil && !errors.Is(err, driver.ErrReleaseExists) {
+		klog.Errorf("Failed to install chart err=%v", err)
+		return err
+	}
+	err = h.addApplicationLabelsToDeployment()
+	if err != nil {
+		h.Uninstall()
+		return err
+	}
+
+	isDepClusterScopedApp := false
+	client, err := versioned.NewForConfig(h.kubeConfig)
+	if err != nil {
+		return err
+	}
+	apps, err := client.AppV1alpha1().Applications().List(h.ctx, metav1.ListOptions{})
+	if err != nil {
+		return err
+	}
+
+	for _, dep := range h.app.Dependencies {
+		if dep.Type == constants.DependencyTypeSystem {
+			continue
+		}
+		for _, app := range apps.Items {
+			if app.Spec.Name == dep.Name && app.Spec.Settings["clusterScoped"] == "true" {
+				isDepClusterScopedApp = true
+				break
+			}
+		}
+
+	}
+	if isDepClusterScopedApp {
+		err = h.addLabelToNamespaceForDependClusterApp()
+		if err != nil {
+			h.Uninstall()
+			return err
+		}
+	}
+	ok, err := h.waitForLaunch()
+
+	if !ok {
+		// install operation has been canceled, so to uninstall it.
+		h.Uninstall()
+		//return context.Canceled
+		return err
+	}
+	klog.Infof("app: %s launched success", h.app.AppName)
+
+	return nil
+}
+
+// NewHelmOps constructs a new helmOps.
+func NewHelmOps(ctx context.Context, kubeConfig *rest.Config, app *appcfg.ApplicationConfig, token string, options Opt) (HelmOpsInterface, error) {
+	actionConfig, settings, err := helm.InitConfig(kubeConfig, app.Namespace)
+	if err != nil {
+		return nil, err
+	}
+
+	client, err := clientset.New(kubeConfig)
+	if err != nil {
+		return nil, err
+	}
+	ops := &HelmOps{
+		ctx:          ctx,
+		kubeConfig:   kubeConfig,
+		app:          app,
+		actionConfig: actionConfig,
+		settings:     settings,
+		token:        token,
+		client:       client,
+		options:      options,
+	}
+	return ops, nil
+}
+
+// addApplicationLabelsToDeployment add application label to deployment or statefulset
+func (h *HelmOps) addApplicationLabelsToDeployment() error {
+	k8s, err := kubernetes.NewForConfig(h.kubeConfig)
+	if err != nil {
+		return err
+	}
+
+	// add namespace to workspace
+	patch := "{\"metadata\": {\"labels\":{\"kubesphere.io/workspace\":\"system-workspace\"}}}"
+	_, err = k8s.CoreV1().Namespaces().Patch(h.ctx, h.app.Namespace,
+		types.MergePatchType, []byte(patch), metav1.PatchOptions{})
+	if err != nil {
+		return err
+	}
+	services := ToEntrancesLabel(h.app.Entrances)
+	ports := ToAppTCPUDPPorts(h.app.Ports)
+
+	tailScale := ToTailScale(h.app.TailScale)
+
+	patchData := map[string]interface{}{
+		"metadata": map[string]interface{}{
+			"labels": map[string]string{
+				constants.ApplicationNameLabel:      h.app.AppName,
+				constants.ApplicationOwnerLabel:     h.app.OwnerName,
+				constants.ApplicationTargetLabel:    h.app.Target,
+				constants.ApplicationRunAsUserLabel: strconv.FormatBool(h.app.RunAsUser),
+			},
+			"annotations": map[string]string{
+				constants.ApplicationIconLabel:    h.app.Icon,
+				constants.ApplicationTitleLabel:   h.app.Title,
+				constants.ApplicationVersionLabel: h.app.Version,
+				constants.ApplicationEntrancesKey: services,
+				constants.ApplicationPortsKey:     ports,
+				constants.ApplicationSourceLabel:  h.options.Source,
+				constants.ApplicationTailScaleKey: tailScale,
+				constants.ApplicationRequiredGPU:  h.app.RequiredGPU,
+			},
+		},
+	}
+
+	patchByte, err := json.Marshal(patchData)
+	if err != nil {
+		return err
+	}
+
+	patch = string(patchByte)
+
+	// TODO: add ownerReferences of user
+	deployment, err := k8s.AppsV1().Deployments(h.app.Namespace).Get(h.ctx, h.app.AppName, metav1.GetOptions{})
+	if err != nil {
+		if apierrors.IsNotFound(err) {
+			return h.tryToAddApplicationLabelsToStatefulSet(k8s, patch)
+		}
+		return err
+	}
+
+	_, err = k8s.AppsV1().Deployments(h.app.Namespace).Patch(h.ctx,
+		deployment.Name,
+		types.MergePatchType,
+		[]byte(patch),
+		metav1.PatchOptions{})
+
+	return err
+}
+
+func (h *HelmOps) tryToAddApplicationLabelsToStatefulSet(k8s *kubernetes.Clientset, patch string) error {
+	statefulSet, err := k8s.AppsV1().StatefulSets(h.app.Namespace).Get(h.ctx, h.app.AppName, metav1.GetOptions{})
+	if err != nil {
+		if apierrors.IsNotFound(err) {
+			return nil
+		}
+		return err
+	}
+
+	_, err = k8s.AppsV1().StatefulSets(h.app.Namespace).Patch(h.ctx,
+		statefulSet.Name,
+		types.MergePatchType,
+		[]byte(patch),
+		metav1.PatchOptions{})
+
+	return err
 }
 
 func (h *HelmOps) status() (*helmrelease.Release, error) {
@@ -104,6 +285,219 @@ func (h *HelmOps) status() (*helmrelease.Release, error) {
 		return nil, err
 	}
 	return status, nil
+}
+
+func (h *HelmOps) addLabelToNamespaceForDependClusterApp() error {
+	k8s, err := kubernetes.NewForConfig(h.kubeConfig)
+	if err != nil {
+		return err
+	}
+
+	labels := map[string]string{
+		constants.ApplicationClusterDep: h.app.AppName,
+	}
+	patchData := map[string]interface{}{"metadata": map[string]map[string]string{"labels": labels}}
+	patchBytes, _ := json.Marshal(patchData)
+	_, err = k8s.CoreV1().Namespaces().Patch(h.ctx, h.app.Namespace,
+		types.MergePatchType, patchBytes, metav1.PatchOptions{})
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+func (h *HelmOps) setValues() (values map[string]interface{}, err error) {
+	values = make(map[string]interface{})
+	values["bfl"] = map[string]interface{}{
+		"username": h.app.OwnerName,
+	}
+	zone, err := h.userZone()
+	if err != nil {
+		klog.Errorf("Failed to find user zone on crd err=%v", err)
+	} else if zone != "" {
+		values["user"] = map[string]interface{}{
+			"zone": zone,
+		}
+	}
+
+	entries := make(map[string]interface{})
+	for i, entrance := range h.app.Entrances {
+		var url string
+		if len(h.app.Entrances) == 1 {
+			url = fmt.Sprintf("%s.%s", h.app.AppID, zone)
+		} else {
+			url = fmt.Sprintf("%s%d.%s", h.app.AppID, i, zone)
+		}
+		entries[entrance.Name] = url
+	}
+
+	values["domain"] = entries
+	userspace := make(map[string]interface{})
+	h.app.Permission = parseAppPermission(h.app.Permission)
+	for _, p := range h.app.Permission {
+		switch perm := p.(type) {
+		case appcfg.AppDataPermission, appcfg.AppCachePermission, appcfg.UserDataPermission:
+
+			// app requests app data permission
+			// set .Values.schedule.nodeName and .Values.userspace.appCache to app
+			// since app data on the bfl's local hostpath, app will schedule to the same node of bfl
+			node, appCachePath, userspacePath, err := h.selectNode()
+			if err != nil {
+				klog.Errorf("Failed select node err=%v", err)
+				return values, err
+			}
+			values["schedule"] = map[string]interface{}{
+				"nodeName": node,
+			}
+
+			// appData = userspacePath + /Data
+			// userData = userspacePath + /Home
+
+			if perm == appcfg.AppCacheRW {
+				userspace["appCache"] = appCachePath
+				if h.options.Source == "devbox" {
+					userspace["appCache"] = filepath.Join(appCachePath, "studio")
+				}
+			}
+			if perm == appcfg.UserDataRW {
+				userspace["userData"] = fmt.Sprintf("%s/Home", userspacePath)
+			}
+			if perm == appcfg.AppDataRW {
+				appData := fmt.Sprintf("%s/Data", userspacePath)
+				userspace["appData"] = appData
+				if h.options.Source == "devbox" {
+					userspace["appData"] = filepath.Join(appData, "studio")
+				}
+			}
+
+		case []appcfg.SysDataPermission:
+			appReg, err := h.registerAppPerm(perm)
+			if err != nil {
+				klog.Errorf("Failed to register err=%v", err)
+				return values, err
+			}
+
+			values["os"] = map[string]interface{}{
+				"appKey":    appReg.Data.AppKey,
+				"appSecret": appReg.Data.AppSecret,
+			}
+		}
+	}
+	values["userspace"] = userspace
+
+	// set service entrance for app that depend on cluster-scoped app
+	type Service struct {
+		EntranceName string
+		Host         string
+		Port         int
+	}
+	var services []Service
+	appClient := versioned.NewForConfigOrDie(h.kubeConfig)
+	apps, err := appClient.AppV1alpha1().Applications().List(context.Background(), metav1.ListOptions{})
+	if err != nil {
+		return values, err
+	}
+
+	clusterScopedAppNamespaces := sets.String{}
+
+	for _, dep := range h.app.Dependencies {
+		// if app is cluster-scoped get its host and port
+		for _, app := range apps.Items {
+			if dep.Type == constants.DependencyTypeApp && app.Spec.Name == dep.Name && app.Spec.Settings["clusterScoped"] == "true" {
+				clusterScopedAppNamespaces.Insert(app.Spec.Namespace)
+				for _, e := range app.Spec.Entrances {
+					services = append(services, Service{
+						Host:         e.Host + "." + app.Spec.Namespace,
+						Port:         int(e.Port),
+						EntranceName: e.Name,
+					})
+				}
+			}
+		}
+	}
+	// set cluster-scoped app's host and port to helm Values
+	dep := make(map[string]interface{})
+	for _, svc := range services {
+		dep[fmt.Sprintf("%s_host", svc.EntranceName)] = svc.Host
+		dep[fmt.Sprintf("%s_port", svc.EntranceName)] = svc.Port
+	}
+	values["dep"] = dep
+
+	kClient, err := kubernetes.NewForConfig(h.kubeConfig)
+	if err != nil {
+		return values, err
+	}
+	svcs := make(map[string]interface{})
+	for ns := range clusterScopedAppNamespaces {
+		servicesList, _ := kClient.CoreV1().Services(ns).List(context.TODO(), metav1.ListOptions{})
+		for _, svc := range servicesList.Items {
+			ports := make([]int32, 0)
+			for _, p := range svc.Spec.Ports {
+				ports = append(ports, p.Port)
+			}
+			svcs[fmt.Sprintf("%s_host", svc.Name)] = fmt.Sprintf("%s.%s", svc.Name, svc.Namespace)
+			svcs[fmt.Sprintf("%s_ports", svc.Name)] = ports
+		}
+	}
+	values["svcs"] = svcs
+	klog.Info("svcs: ", svcs)
+
+	var arch string
+	nodes, err := kClient.CoreV1().Nodes().List(context.TODO(), metav1.ListOptions{})
+	if err != nil {
+		return values, err
+	}
+	for _, node := range nodes.Items {
+		arch = node.Labels["kubernetes.io/arch"]
+		break
+	}
+	values["cluster"] = map[string]interface{}{
+		"arch": arch,
+	}
+	gpuType, err := utils.FindGpuTypeFromNodes(nodes)
+	if err != nil {
+		klog.Errorf("Failed to get gpuType err=%v", err)
+		return values, err
+	}
+	values["GPU"] = map[string]interface{}{
+		"Type": gpuType,
+		"Cuda": os.Getenv("CUDA_VERSION"),
+	}
+
+	values["gpu"] = gpuType
+
+	if h.app.OIDC.Enabled {
+		err = h.createOIDCClient(values, zone, h.app.Namespace)
+		if err != nil {
+			klog.Errorf("Failed to create OIDCClient err=%v", err)
+			return values, err
+		}
+	}
+
+	sharedLibPath := os.Getenv("SHARED_LIB_PATH")
+	values["sharedlib"] = sharedLibPath
+
+	admin, err := kubesphere.GetAdminUsername(context.TODO(), h.kubeConfig)
+	if err != nil {
+		return values, err
+	}
+	values["admin"] = admin
+
+	isAdmin, err := kubesphere.IsAdmin(context.TODO(), h.kubeConfig, h.app.OwnerName)
+	if err != nil {
+		return values, err
+	}
+	values["isAdmin"] = isAdmin
+
+	rootPath := userspacev1.DefaultRootPath
+	if os.Getenv(userspacev1.OlaresRootPath) != "" {
+		rootPath = os.Getenv(userspacev1.OlaresRootPath)
+	}
+	values["rootPath"] = rootPath
+
+	values["downloadCdnURL"] = os.Getenv("DOWNLOAD_CDN_URL")
+
+	return values, err
 }
 
 func (h *HelmOps) userZone() (string, error) {
@@ -714,7 +1108,7 @@ func (h *HelmOps) isStartUp() (bool, error) {
 		creationTime := pod.GetCreationTimestamp()
 		pendingDuration := time.Since(creationTime.Time)
 
-		if pod.Status.Phase == corev1.PodPending && pendingDuration > time.Minute*10 {
+		if pod.Status.Phase == corev1.PodPending && pendingDuration > time.Minute*2 {
 			return false, errcode.ErrPodPending
 		}
 		totalContainers := len(pod.Spec.Containers)
@@ -848,10 +1242,69 @@ func parseAppPermission(data []appcfg.AppPermission) []appcfg.AppPermission {
 }
 
 func (h *HelmOps) Install() error {
-	if h.installStrategy == nil {
-		h.installStrategy = NewInstallV1(h)
+	var err error
+	values, err := h.setValues()
+	if err != nil {
+		klog.Errorf("set values err %v", err)
+		return err
 	}
-	return h.installStrategy.Install()
+	namespace := fmt.Sprintf("%s-%s", "user-system", h.app.OwnerName)
+	if err := tapr.Apply(h.app.Middleware, h.kubeConfig, h.app.AppName, h.app.Namespace,
+		namespace, h.token, h.app.ChartsName, h.app.OwnerName, values); err != nil {
+		klog.Errorf("Failed to apply middleware err=%v", err)
+		return err
+	}
+	err = h.install(values)
+	if err != nil && !errors.Is(err, driver.ErrReleaseExists) {
+		klog.Errorf("Failed to install chart err=%v", err)
+		h.Uninstall()
+		return err
+	}
+	err = h.addApplicationLabelsToDeployment()
+	if err != nil {
+		h.Uninstall()
+		return err
+	}
+
+	isDepClusterScopedApp := false
+	client, err := versioned.NewForConfig(h.kubeConfig)
+	if err != nil {
+		return err
+	}
+	apps, err := client.AppV1alpha1().Applications().List(h.ctx, metav1.ListOptions{})
+	if err != nil {
+		return err
+	}
+
+	for _, dep := range h.app.Dependencies {
+		if dep.Type == constants.DependencyTypeSystem {
+			continue
+		}
+		for _, app := range apps.Items {
+			if app.Spec.Name == dep.Name && app.Spec.Settings["clusterScoped"] == "true" {
+				isDepClusterScopedApp = true
+				break
+			}
+		}
+
+	}
+	if isDepClusterScopedApp {
+		err = h.addLabelToNamespaceForDependClusterApp()
+		if err != nil {
+			h.Uninstall()
+			return err
+		}
+	}
+
+	ok, err := h.waitForStartUp()
+	if err != nil && errors.Is(err, errcode.ErrPodPending) {
+		return err
+	}
+	if !ok {
+		h.Uninstall()
+		return err
+	}
+	return nil
 }
 
 func (h *HelmOps) isPending() error {
@@ -889,48 +1342,4 @@ func (h *HelmOps) WaitForLaunch() (bool, error) {
 			return false, h.ctx.Err()
 		}
 	}
-}
-
-type InstallStrategy interface {
-	Install() error
-}
-
-type InstallVersion string
-
-const (
-	InstallVersionV1 InstallVersion = "v1"
-	InstallVersionV2 InstallVersion = "v2"
-)
-
-func NewHelmOpsWithVersion(ctx context.Context, kubeConfig *rest.Config, app *appcfg.ApplicationConfig, token string, options Opt, version InstallVersion) (HelmOpsInterface, error) {
-	actionConfig, settings, err := helm.InitConfig(kubeConfig, app.Namespace)
-	if err != nil {
-		return nil, err
-	}
-
-	client, err := clientset.New(kubeConfig)
-	if err != nil {
-		return nil, err
-	}
-	ops := &HelmOps{
-		ctx:          ctx,
-		kubeConfig:   kubeConfig,
-		app:          app,
-		actionConfig: actionConfig,
-		settings:     settings,
-		token:        token,
-		client:       client,
-		options:      options,
-	}
-
-	switch version {
-	case InstallVersionV1:
-		ops.setInstallStrategy(NewInstallV1(ops))
-	case InstallVersionV2:
-		ops.setInstallStrategy(NewInstallV2(ops))
-	default:
-		ops.setInstallStrategy(NewInstallV1(ops))
-	}
-
-	return ops, nil
 }
