+++ pkg/appstate/initializing_app.go
@@ -59,7 +59,7 @@ func (p *InitializingApp) Exec(ctx context.Context) (StatefulInProgressApp, erro
 
 	opCtx, cancel := context.WithCancel(context.Background())
 
-	ops, err := appinstaller.NewHelmOpsWithVersion(opCtx, kubeConfig, appCfg, token, appinstaller.Opt{Source: p.manager.Spec.Source}, "v1")
+	ops, err := appinstaller.NewHelmOps(opCtx, kubeConfig, appCfg, token, appinstaller.Opt{Source: p.manager.Spec.Source})
 	if err != nil {
 		klog.Errorf("make helm ops failed %v", err)
 		cancel()
