+++ pkg/appstate/types.go
@@ -87,7 +87,7 @@ func (p *baseStatefulApp) forceDeleteApp(ctx context.Context) error {
 		klog.Errorf("get kube config failed %v", err)
 		return err
 	}
-	ops, err := appinstaller.NewHelmOpsWithVersion(ctx, kubeConfig, appCfg, token, appinstaller.Opt{}, "v1")
+	ops, err := appinstaller.NewHelmOps(ctx, kubeConfig, appCfg, token, appinstaller.Opt{})
 	if err != nil {
 		klog.Errorf("make helm ops failed %v", err)
 		return err
